Adapt openqa-scripts to new versions of openQA-client
Recent versions of openQA-client (at least openQA-client-5.1758036156,
but perhaps some older ones as well) have broken the scripts included in
this repository. Last known openQA-client version which was compatible
was openQA-client-5.1749827588.

Primary issue is that when attempting to use `OpenQA::CLI::api` directly
as these scripts were doing, Perl would throw an `Undefined subroutine`
exception pointing to `&OpenQA::CLI::get_opt`. But the subroutine is
indeed declared in `OpenQA::CLI`.

This can be solved by calling the client api from the super class
`OpenQA::CLI` instead.

Another issue noticed is that now `OpenQA::CLI` will search for some
YAML specs which are usually located in:

* /usr/share/openqa/public/openqa-cli.yaml
* /usr/share/openqa/client/openqa-cli.yaml

But the way `OpenQA::CLI` searches for those is based on the directory
where the script was called from, making it necessary for these script
to reside in the same directory as `openqa-cli` (usually this would be
`/usr/share/openqa/script`). As there is no way to work around this
issue at the moment, documentation has been updated with an
**INSTALLATION** section recommending installing the scripts in that
directory.